+++ example/ionic-angular-v7/src/app/app.module.ts
@@ -23,6 +23,12 @@ Sentry.init(
     // Whether SDK should be enabled or not
     enabled: true,
     // Use the tracing integration to see traces and add performance monitoring
+    _experiments: {
+      enableMetrics: true,
+    },
+    beforeSendMetric: (metric) => {
+      return metric;
+    },
     integrations: [
       Sentry.browserTracingIntegration(),
       Sentry.replayIntegration({